% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_love.R
\name{in_love}
\alias{in_love}
\title{Friday I'm In Love by the Cure}
\usage{
in_love(best_day, other_day)
}
\arguments{
\item{best_day}{a character that should literally only be Friday}

\item{other_day}{a character that can be any day of the week}
}
\value{

}
\description{
Friday I'm In Love by the Cure
}
