% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad_lib_1.R
\name{mad_lib_1}
\alias{mad_lib_1}
\title{Make a very simple madlib!}
\usage{
mad_lib_1(name, food, adjective)
}
\arguments{
\item{name}{is a character of the name you want to use}

\item{food}{is a character of a type of food}

\item{adjective}{is a character that describes the way the individual ate}
}
\value{

}
\description{
Make a very simple madlib!
}
