% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_me.R
\name{name_birthday}
\alias{name_birthday}
\title{Tells you a persons name and their birthday}
\usage{
name_birthday(name, day, month)
}
\arguments{
\item{name}{a character containing the input name}

\item{day}{a number containing the input day of the month you were born}

\item{month}{a character containing the month of your birthday}
}
\value{

}
\description{
Contains one function, name_birthday, that prints a statement about the user
}
