% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_horrible_test.R
\name{theme_horrible_test}
\alias{theme_horrible_test}
\title{Horrendous test theme}
\usage{
theme_horrible_test()
}
\value{

}
\description{
Description: We made this theme as an example of how we can make custom themes to save for future use.
}
